Start script now waits for Kafka process to launch befor starting schema registry. Container now exports /logs, /kafka-logs, and /zookeeper.

+++ scripts/confluent-start.sh
@@ -1,25 +1,53 @@
 #!/bin/bash
 
+mkdir /logs &> /dev/null
+
 cd /confluent-2.0.0
 
-IP=$(cat /etc/hosts | head -n1 | awk '{print $1}')
+_IP=$(cat /etc/hosts | head -n1 | awk '{print $1}')
+
+_KAFKA_DATA_DIR=/kafka-logs
+
+_ZK_DATA_DIR=/zookeeper
 
 kafka_cfg=./etc/kafka/server.properties
 kafka_cfg_tpl=./etc/kafka/server.properties.tpl
 
-cp $kafka_cfg $kafka_cfg_tpl
+if [ ! -f "$kafka_cfg_tpl" ]; then
+	cp $kafka_cfg $kafka_cfg_tpl
+fi
+
+zk_cfg=./etc/kafka/zookeeper.properties
+zk_cfg_tpl=./etc/kafka/zookeeper.properties.tpl
+
+if [ ! -f "$zk_cfg_tpl" ]; then
+	cp $zk_cfg $zk_cfg_tpl
+fi
 
 cat $kafka_cfg_tpl | sed \
-	-e "s|#advertised.host.name=<hostname routable by clients>|advertised.host.name=${KAFKA_ADVERTISED_HOST_NAME:-$IP}|g" \
+	-e "s|#advertised.host.name=<hostname routable by clients>|advertised.host.name=${KAFKA_ADVERTISED_HOST_NAME:-$_IP}|g" \
+	-e "s|log.dirs=/tmp/kafka-logs|log.dirs=${KAFKA_DATA_DIR:-$_KAFKA_DATA_DIR}|g" \
 	> $kafka_cfg
 
-rm $kafka_cfg_tpl
+cat $zk_cfg_tpl | sed \
+	-e "s|dataDir=/tmp/zookeeper|dataDir=${ZK_DATA_DIR:-$_ZK_DATA_DIR}|g" \
+	> $zk_cfg
 
-./bin/zookeeper-server-start ./etc/kafka/zookeeper.properties &
+./bin/zookeeper-server-start ./etc/kafka/zookeeper.properties | tee /logs/zookeeper.log &
 sleep 5
-./bin/kafka-server-start $kafka_cfg &
-sleep 5
-./bin/schema-registry-start ./etc/schema-registry/schema-registry.properties &
+
+./bin/kafka-server-start $kafka_cfg | tee /logs/kafka.log &
+sleep 1
+
+echo "Waiting for Kafka to start..."
+timeout 15 grep -q 'Starting metrics collection from monitored broker' <(tail -F /logs/kafka.log)
+
+if [ "$?" -eq 124 ]; then
+	echo "error: could not start Kafka server, check logs." >&2
+	exit 1
+fi
+
+./bin/schema-registry-start ./etc/schema-registry/schema-registry.properties | tee /logs/schema-registry.log &
 
 wait
